fix[User]: Move to correct package.
