+++ src/main/java/fr/oukilson/backend/dto/user/UserCreationDTO.java
@@ -1,4 +1,4 @@
-package fr.oukilson.backend.dto;
+package fr.oukilson.backend.dto.user;
 
 import lombok.*;
 import javax.validation.constraints.NotNull;
