+++ src/main/java/fr/oukilson/backend/dto/user/UserDTO.java
@@ -1,0 +1,12 @@
+package fr.oukilson.backend.dto.user;
+
+import lombok.*;
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserDTO {
+    private String nickname;
+    private List<UserNameDTO> friendList;
+}
